pyproject.toml
[tool.poetry]
name = "cat-discord-bot"
version = "0.2.1"
description = ""
authors = ["Walfask"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
discord = "^1.7.3"
python-dotenv = "^0.20.0"
aiocron = "^1.8"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"
pre-commit = "^2.20.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
